+++ CHANGELOG.md
@@ -3,13 +3,13 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 This CHANGELOG follows the format listed at [Keep A Changelog](http://keepachangelog.com/)
 
-## Unreleased][unreleased]
+## Unreleased
 
 ### Added
 - Timeout option.
+- Basic chronos check.
 
 ## 0.0.1 - 2015-05-21
 
 ### Added
 - initial release
-